refactor(hook_handlers): split mod.rs to clear the LOC gate
hook_handlers/mod.rs grew to 1537 lines, over scripts/loc-gate.sh's
1500 line limit, failing CI for every PR against main regardless of
what it touches. The gate's own policy says to shrink the allowlist
and never extend it, so this splits the file into submodules instead
of allowlisting it, mirroring the existing search_rewrite extraction
from #660.

Three new files take over most of the content. file_rewrite.rs owns
file-read (cat, head, tail, Get-Content) rewriting, compound-command
wrapping, and the rewrite_candidate dispatch every rewrite entry point
funnels through. redirect.rs owns the Read, Grep and Glob redirect
decision logic. codex.rs owns the Codex CLI PreToolUse handlers and
SessionStart guidance.

mod.rs now holds only shared utilities and dispatch entry points,
dropping from 1537 to 388 lines. Functions needed only by the test
module (via a wildcard super import) are re-exported behind a test-only
cfg gate so production builds don't carry unused-import warnings for
names only test code references by path.

All existing hook_handlers tests pass unchanged; this is pure code
motion with no behavior change.

Fixes #966
